+++ tests/Dummy/PhpDto.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Anteris\Tests\DataTransferObjectFactory\Dummy;
+
+class PhpDto
+{
+    public static bool $booted = false;
+    
+    public string $firstName;
+    public string $lastName;
+    public null | bool $active;
+}